修改結構,新增會員管理,會員訂單查詢

+++ src/main/java/com/vendingmachine/backend/vo/ProductVo.java
@@ -27,7 +27,7 @@ public class ProductVo extends JSGridFilter{
 	
 	private Integer priceEnd;
 	
-	private Integer cost;
+	private Double cost;
 	
 	private Integer costStart;
